pyproject.toml
[tool.poetry]
name = "nonebot-plugin-setu-now"
version = "0.1.0"
description = "另一个色图插件"
authors = ["kexue"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/kexue-z/nonebot-plugin-setu-now"

[tool.poetry.dependencies]
python = "^3.8"
httpx = ">=0.18.0,<1.0.0"
webdav4 = "^0.9.3"
nonebot2 = "^2.0.0-beta.1"
nonebot-adapter-onebot = "^2.0.0-beta.1"
pydantic = "^1.5.0"
nonebot-plugin-apscheduler = "^0.1.2"
Pillow = ">=8"

[tool.poetry.dev-dependencies]
black = {version = "^21.12b0", allow-prereleases = true}
flake8 = "^4.0.1"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
